# src/lumio-web/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Lumio API",
    "description": "Offline digitale nalatenschap beheer — Lumio",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/eigenaar": {
      "get": {
        "tags": [
          "Eigenaar"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EigenaarResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EigenaarResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EigenaarResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Eigenaar"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EigenaarUpsertRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EigenaarUpsertRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EigenaarUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EigenaarResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EigenaarResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EigenaarResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Eigenaar"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EigenaarUpsertRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EigenaarUpsertRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EigenaarUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EigenaarResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EigenaarResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EigenaarResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/eigenaar/foto": {
      "get": {
        "tags": [
          "Eigenaar"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Eigenaar"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "Bestand"
                ],
                "type": "object",
                "properties": {
                  "Bestand": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "Bestand": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Eigenaar"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/shamir/genereer": {
      "post": {
        "tags": [
          "Shamir"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenereerSharesRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GenereerSharesRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GenereerSharesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GenereerSharesResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenereerSharesResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenereerSharesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shamir/reconstrueer-en-ontgrendel": {
      "post": {
        "tags": [
          "Shamir"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReconstrueerRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReconstrueerRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReconstrueerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BurgerlijkeStaat": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "EigenaarResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "voornaam": {
            "type": "string",
            "nullable": true
          },
          "achternaam": {
            "type": "string",
            "nullable": true
          },
          "tussenvoegsel": {
            "type": "string",
            "nullable": true
          },
          "geboortedatum": {
            "type": "string",
            "format": "date"
          },
          "bsn": {
            "type": "string",
            "nullable": true
          },
          "adres": {
            "type": "string",
            "nullable": true
          },
          "postcode": {
            "type": "string",
            "nullable": true
          },
          "woonplaats": {
            "type": "string",
            "nullable": true
          },
          "telefoon": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "notaris": {
            "type": "string",
            "nullable": true
          },
          "notarisKantoor": {
            "type": "string",
            "nullable": true
          },
          "notarisTelefoon": {
            "type": "string",
            "nullable": true
          },
          "notarisEmail": {
            "type": "string",
            "nullable": true
          },
          "notarisAdres": {
            "type": "string",
            "nullable": true
          },
          "notarisPostcode": {
            "type": "string",
            "nullable": true
          },
          "notarisPlaats": {
            "type": "string",
            "nullable": true
          },
          "burgerlijkeStaat": {
            "$ref": "#/components/schemas/BurgerlijkeStaat"
          },
          "huwelijksVoorwaarden": {
            "$ref": "#/components/schemas/HuwelijksVoorwaarden"
          },
          "datumHuwelijk": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "legitimatieSoort": {
            "$ref": "#/components/schemas/LegitimatieSoort"
          },
          "legitimatieNummer": {
            "type": "string",
            "nullable": true
          },
          "legitimatieDatumAfgifte": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "legitimatieGeldigTot": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "heeftProfielFoto": {
            "type": "boolean"
          },
          "aangemaaktOp": {
            "type": "string",
            "format": "date-time"
          },
          "gewijzigdOp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "EigenaarUpsertRequest": {
        "type": "object",
        "properties": {
          "voornaam": {
            "type": "string",
            "nullable": true
          },
          "achternaam": {
            "type": "string",
            "nullable": true
          },
          "tussenvoegsel": {
            "type": "string",
            "nullable": true
          },
          "geboortedatum": {
            "type": "string",
            "format": "date"
          },
          "bsn": {
            "type": "string",
            "nullable": true
          },
          "adres": {
            "type": "string",
            "nullable": true
          },
          "postcode": {
            "type": "string",
            "nullable": true
          },
          "woonplaats": {
            "type": "string",
            "nullable": true
          },
          "telefoon": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "notaris": {
            "type": "string",
            "nullable": true
          },
          "notarisKantoor": {
            "type": "string",
            "nullable": true
          },
          "notarisTelefoon": {
            "type": "string",
            "nullable": true
          },
          "notarisEmail": {
            "type": "string",
            "nullable": true
          },
          "notarisAdres": {
            "type": "string",
            "nullable": true
          },
          "notarisPostcode": {
            "type": "string",
            "nullable": true
          },
          "notarisPlaats": {
            "type": "string",
            "nullable": true
          },
          "burgerlijkeStaat": {
            "$ref": "#/components/schemas/BurgerlijkeStaat"
          },
          "huwelijksVoorwaarden": {
            "$ref": "#/components/schemas/HuwelijksVoorwaarden"
          },
          "datumHuwelijk": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "legitimatieSoort": {
            "$ref": "#/components/schemas/LegitimatieSoort"
          },
          "legitimatieNummer": {
            "type": "string",
            "nullable": true
          },
          "legitimatieDatumAfgifte": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "legitimatieGeldigTot": {
            "type": "string",
            "format": "date",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GenereerSharesRequest": {
        "type": "object",
        "properties": {
          "wachtwoord": {
            "type": "string",
            "nullable": true
          },
          "aantalDelen": {
            "type": "integer",
            "format": "int32"
          },
          "drempel": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "GenereerSharesResponse": {
        "type": "object",
        "properties": {
          "delen": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShareInfo"
            },
            "nullable": true
          },
          "drempel": {
            "type": "integer",
            "format": "int32"
          },
          "totaalAantalDelen": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "HuwelijksVoorwaarden": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "LegitimatieSoort": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "ReconstrueerRequest": {
        "type": "object",
        "properties": {
          "delen": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ShareInfo": {
        "type": "object",
        "properties": {
          "index": {
            "type": "integer",
            "format": "int32"
          },
          "waarde": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}